admin/apps/products/producer.py
import pika
from decouple import config


URL = config("URL")
params = pika.URLParameters(url=URL)
connection = pika.BlockingConnection(params)
channel = connection.channel()


def publish():
    channel.basic_publish(exchange='', routing_key='main',
                          body='Hi! desde Admin/Producer')
