admin/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
django-mysql==4.10.0
djangorestframework==3.14.0
mysqlclient==2.1.1
pika==1.3.2
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
